requirements.txt
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
asyncmy==0.2.9
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
camel-converter==3.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
exceptiongroup==1.2.0
executing==2.0.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
meilisearch==0.28.4
meilisearch-python-sdk==2.2.3
meilisync==0.1.2
motor==3.3.2
mysql-connector-python==8.2.0
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.21.12
psutil==5.9.7
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
streamlit==1.29.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
traitlets==5.14.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
zipp==3.17.0
